Make navbar responsive.
